refactor: Move socket functions to buddySocket file
